move alert dict casting fncs to utils module and bump version. figures and pubsub modules both use these fncs.
